requirements.txt
aiogram>=3.5.0,<4.0.0
aiosqlite>=0.19.0
python-dotenv>=1.0.1
apscheduler>=3.10.4
pytz>=2024.1

aiogram>=3.7
python-dotenv
fastapi
uvicorn
httpx>=0.24.0
